test: move tests t orest_jobrunner.py
