+++ tests/test_jobrunner.py
@@ -1,6 +1,10 @@
 from pathlib import Path
 from types import SimpleNamespace
 
+from click.testing import CliRunner
+
+from chemsmart.cli.run import run
+from chemsmart.cli.sub import sub
 from chemsmart.jobs.gaussian.runner import (
     FakeGaussianJobRunner,
     GaussianJobRunner,
@@ -8,6 +12,7 @@ from chemsmart.jobs.gaussian.runner import (
 from chemsmart.jobs.iterate.runner import IterateJobRunner
 from chemsmart.jobs.orca.runner import FakeORCAJobRunner, ORCAJobRunner
 from chemsmart.jobs.runner import JobRunner
+from chemsmart.settings.server import Server
 
 
 class DummyORCAJob:
@@ -149,3 +154,176 @@ class TestJobRunnerSelection:
         assert Path(runner.job_gbwfile).name == "orca_opt_fake.gbw"
         assert Path(runner.job_errfile).name == "orca_opt_fake.err"
         assert Path(runner.job_outputfile).name == "orca_opt_fake.out"
+
+
+def _write_gaussian_project(tmp_path):
+    config_root = tmp_path / "chemsmart_cfg"
+    gaussian_cfg = config_root / "gaussian"
+    gaussian_cfg.mkdir(parents=True)
+    (gaussian_cfg / "test.yaml").write_text(
+        "gas:\n  functional: B3LYP\n  basis: def2-SVP\n"
+        "solv:\n  functional: B3LYP\n  basis: def2-SVP\n"
+        "  solvent_model: smd\n  solvent_id: water\n"
+    )
+    return config_root
+
+
+class TestScratchCLIDefaults:
+    """Omitted --scratch should leave program defaults via from_job."""
+
+    def test_run_omitted_scratch_leaves_none_for_program_default(
+        self, tmp_path, monkeypatch, single_molecule_xyz_file
+    ):
+        monkeypatch.setenv(
+            "CHEMSMART_CONFIG_DIR", str(_write_gaussian_project(tmp_path))
+        )
+        observed = {"scratch_arg": "unset"}
+
+        def _from_job(cls, job, server, scratch=None, fake=False, **kwargs):
+            observed["scratch_arg"] = scratch
+            return type("R", (), {"scratch": scratch})()
+
+        monkeypatch.setattr(
+            "chemsmart.jobs.runner.JobRunner.from_job",
+            classmethod(_from_job),
+        )
+        monkeypatch.setattr("chemsmart.jobs.job.Job.run", lambda self: None)
+
+        result = CliRunner().invoke(
+            run,
+            [
+                "--fake",
+                "gaussian",
+                "-p",
+                "test",
+                "-f",
+                single_molecule_xyz_file,
+                "-c",
+                "0",
+                "-m",
+                "1",
+                "opt",
+            ],
+        )
+        assert result.exit_code == 0, result.output
+        assert observed["scratch_arg"] is None
+
+    def test_sub_omitted_scratch_does_not_force_no_scratch(
+        self, tmp_path, monkeypatch, single_molecule_xyz_file
+    ):
+        monkeypatch.setenv(
+            "CHEMSMART_CONFIG_DIR", str(_write_gaussian_project(tmp_path))
+        )
+        fake_server = Server(name="dummy")
+        captured = {"cli_args": None}
+        fake_server.submit = (
+            lambda job, test=False, cli_args=None, **kw: captured.update(
+                cli_args=cli_args
+            )
+        )
+        monkeypatch.setattr(
+            "chemsmart.settings.server.Server.from_servername",
+            lambda _name: fake_server,
+        )
+
+        result = CliRunner().invoke(
+            sub,
+            [
+                "--test",
+                "--server",
+                "dummy",
+                "gaussian",
+                "-p",
+                "test",
+                "-f",
+                single_molecule_xyz_file,
+                "-c",
+                "0",
+                "-m",
+                "1",
+                "opt",
+            ],
+        )
+        assert result.exit_code == 0, result.output
+        assert "--no-scratch" not in captured["cli_args"]
+
+    def test_run_explicit_scratch_is_not_silently_disabled(
+        self, tmp_path, monkeypatch, single_molecule_xyz_file
+    ):
+        """Placeholder JobRunner must not clear --scratch before from_job."""
+        monkeypatch.setenv(
+            "CHEMSMART_CONFIG_DIR", str(_write_gaussian_project(tmp_path))
+        )
+        observed = {"scratch_arg": "unset"}
+
+        def _from_job(cls, job, server, scratch=None, fake=False, **kwargs):
+            observed["scratch_arg"] = scratch
+            return type("R", (), {"scratch": scratch})()
+
+        monkeypatch.setattr(
+            "chemsmart.jobs.runner.JobRunner.from_job",
+            classmethod(_from_job),
+        )
+        monkeypatch.setattr("chemsmart.jobs.job.Job.run", lambda self: None)
+
+        result = CliRunner().invoke(
+            run,
+            [
+                "--fake",
+                "--scratch",
+                "gaussian",
+                "-p",
+                "test",
+                "-f",
+                single_molecule_xyz_file,
+                "-c",
+                "0",
+                "-m",
+                "1",
+                "opt",
+            ],
+        )
+        assert result.exit_code == 0, result.output
+        assert observed["scratch_arg"] is True
+
+    def test_sub_explicit_scratch_keeps_jobrunner_scratch_true(
+        self, tmp_path, monkeypatch, single_molecule_xyz_file
+    ):
+        monkeypatch.setenv(
+            "CHEMSMART_CONFIG_DIR", str(_write_gaussian_project(tmp_path))
+        )
+        fake_server = Server(name="dummy")
+        captured = {"job": None, "cli_args": None}
+
+        def _submit(job, test=False, cli_args=None, **kw):
+            captured["job"] = job
+            captured["cli_args"] = cli_args
+
+        fake_server.submit = _submit
+        monkeypatch.setattr(
+            "chemsmart.settings.server.Server.from_servername",
+            lambda _name: fake_server,
+        )
+
+        result = CliRunner().invoke(
+            sub,
+            [
+                "--test",
+                "--scratch",
+                "--server",
+                "dummy",
+                "gaussian",
+                "-p",
+                "test",
+                "-f",
+                single_molecule_xyz_file,
+                "-c",
+                "0",
+                "-m",
+                "1",
+                "opt",
+            ],
+        )
+        assert result.exit_code == 0, result.output
+        assert "--scratch" in captured["cli_args"]
+        assert captured["job"].jobrunner.scratch is True
