+++ tests/test_server.py
@@ -1,10 +1,6 @@
 import os
 from io import StringIO
 
-from click.testing import CliRunner
-
-from chemsmart.cli.run import run
-from chemsmart.cli.sub import sub
 from chemsmart.settings.executable import GaussianExecutable, ORCAExecutable
 from chemsmart.settings.server import Server
 from chemsmart.settings.submitters import PBSSubmitter, SLURMSubmitter
@@ -115,176 +111,3 @@ export LD_LIBRARY_PATH=~/programs/openmpi-4.1.6/build/lib:$LD_LIBRARY_PATH
         buffer = StringIO()
         submitter._write_scheduler_options(buffer)
         assert "#PBS -m abe\n" in buffer.getvalue()
-
-
-def _write_gaussian_project(tmp_path):
-    config_root = tmp_path / "chemsmart_cfg"
-    gaussian_cfg = config_root / "gaussian"
-    gaussian_cfg.mkdir(parents=True)
-    (gaussian_cfg / "test.yaml").write_text(
-        "gas:\n  functional: B3LYP\n  basis: def2-SVP\n"
-        "solv:\n  functional: B3LYP\n  basis: def2-SVP\n"
-        "  solvent_model: smd\n  solvent_id: water\n"
-    )
-    return config_root
-
-
-class TestScratchCLIDefaults:
-    """Omitted --scratch should leave program defaults via from_job."""
-
-    def test_run_omitted_scratch_leaves_none_for_program_default(
-        self, tmp_path, monkeypatch, single_molecule_xyz_file
-    ):
-        monkeypatch.setenv(
-            "CHEMSMART_CONFIG_DIR", str(_write_gaussian_project(tmp_path))
-        )
-        observed = {"scratch_arg": "unset"}
-
-        def _from_job(cls, job, server, scratch=None, fake=False, **kwargs):
-            observed["scratch_arg"] = scratch
-            return type("R", (), {"scratch": scratch})()
-
-        monkeypatch.setattr(
-            "chemsmart.jobs.runner.JobRunner.from_job",
-            classmethod(_from_job),
-        )
-        monkeypatch.setattr("chemsmart.jobs.job.Job.run", lambda self: None)
-
-        result = CliRunner().invoke(
-            run,
-            [
-                "--fake",
-                "gaussian",
-                "-p",
-                "test",
-                "-f",
-                single_molecule_xyz_file,
-                "-c",
-                "0",
-                "-m",
-                "1",
-                "opt",
-            ],
-        )
-        assert result.exit_code == 0, result.output
-        assert observed["scratch_arg"] is None
-
-    def test_sub_omitted_scratch_does_not_force_no_scratch(
-        self, tmp_path, monkeypatch, single_molecule_xyz_file
-    ):
-        monkeypatch.setenv(
-            "CHEMSMART_CONFIG_DIR", str(_write_gaussian_project(tmp_path))
-        )
-        fake_server = Server(name="dummy")
-        captured = {"cli_args": None}
-        fake_server.submit = (
-            lambda job, test=False, cli_args=None, **kw: captured.update(
-                cli_args=cli_args
-            )
-        )
-        monkeypatch.setattr(
-            "chemsmart.settings.server.Server.from_servername",
-            lambda _name: fake_server,
-        )
-
-        result = CliRunner().invoke(
-            sub,
-            [
-                "--test",
-                "--server",
-                "dummy",
-                "gaussian",
-                "-p",
-                "test",
-                "-f",
-                single_molecule_xyz_file,
-                "-c",
-                "0",
-                "-m",
-                "1",
-                "opt",
-            ],
-        )
-        assert result.exit_code == 0, result.output
-        assert "--no-scratch" not in captured["cli_args"]
-
-    def test_run_explicit_scratch_is_not_silently_disabled(
-        self, tmp_path, monkeypatch, single_molecule_xyz_file
-    ):
-        """Placeholder JobRunner must not clear --scratch before from_job."""
-        monkeypatch.setenv(
-            "CHEMSMART_CONFIG_DIR", str(_write_gaussian_project(tmp_path))
-        )
-        observed = {"scratch_arg": "unset"}
-
-        def _from_job(cls, job, server, scratch=None, fake=False, **kwargs):
-            observed["scratch_arg"] = scratch
-            return type("R", (), {"scratch": scratch})()
-
-        monkeypatch.setattr(
-            "chemsmart.jobs.runner.JobRunner.from_job",
-            classmethod(_from_job),
-        )
-        monkeypatch.setattr("chemsmart.jobs.job.Job.run", lambda self: None)
-
-        result = CliRunner().invoke(
-            run,
-            [
-                "--fake",
-                "--scratch",
-                "gaussian",
-                "-p",
-                "test",
-                "-f",
-                single_molecule_xyz_file,
-                "-c",
-                "0",
-                "-m",
-                "1",
-                "opt",
-            ],
-        )
-        assert result.exit_code == 0, result.output
-        assert observed["scratch_arg"] is True
-
-    def test_sub_explicit_scratch_keeps_jobrunner_scratch_true(
-        self, tmp_path, monkeypatch, single_molecule_xyz_file
-    ):
-        monkeypatch.setenv(
-            "CHEMSMART_CONFIG_DIR", str(_write_gaussian_project(tmp_path))
-        )
-        fake_server = Server(name="dummy")
-        captured = {"job": None, "cli_args": None}
-
-        def _submit(job, test=False, cli_args=None, **kw):
-            captured["job"] = job
-            captured["cli_args"] = cli_args
-
-        fake_server.submit = _submit
-        monkeypatch.setattr(
-            "chemsmart.settings.server.Server.from_servername",
-            lambda _name: fake_server,
-        )
-
-        result = CliRunner().invoke(
-            sub,
-            [
-                "--test",
-                "--scratch",
-                "--server",
-                "dummy",
-                "gaussian",
-                "-p",
-                "test",
-                "-f",
-                single_molecule_xyz_file,
-                "-c",
-                "0",
-                "-m",
-                "1",
-                "opt",
-            ],
-        )
-        assert result.exit_code == 0, result.output
-        assert "--scratch" in captured["cli_args"]
-        assert captured["job"].jobrunner.scratch is True
